requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
sqlalchemy>=2.0.20
alembic>=1.12.0
mysql-connector-python>=8.1.0
python-dotenv>=1.0.0
pydantic>=2.3.0
pydantic-settings>=2.0.3
httpx>=0.25.0
cachetools>=5.3.0
loguru>=0.7.0
